Include -Wall compiler flag, and fixed some warnings

+++ src/db.cc
@@ -174,7 +174,7 @@ bool DB::TableSchema(const std::string& table_name, std::string* serialized_sche
 }
 
 bool DB::DropTable(const std::string& table_name) {
-    for (int i = 0; i < table_handles_.size(); i++) {
+    for (size_t i = 0; i < table_handles_.size(); i++) {
         rocksdb::DB* h = table_handles_.at(i);
         if (h->GetName().compare(GetTablePath(table_name)) == 0) {
             delete h;
